Create rule30_to_midi_enhanced.py
This is my try at using Wolframs Cellular Rule30 Automata to create quick random midi files with your favorite daw or plugin.  This is super helpful for quick ideas and is open source and this is just the start i really will try to add all 350 rules of cellular automata and my goal is just to make 4 to 8 bar midi's but as you can see this does quite well with very lengthy ones. I wanted to make this free as my goal is to make a very good midi creator that is open source and as good as anything outthere